requirements.txt
anyio==4.10.0
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.1.1
python-telegram-bot==22.3
sniffio==1.3.1
typing_extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0
fastapi==0.116.1
pydantic==2.11.7
pydantic_core==2.33.2
starlette==0.47.2
charset-normalizer==3.4.2
requests==2.32.4
urllib3==2.5.0
click==8.2.1
colorama==0.4.6
uvicorn==0.35.0